Refactor session.py
